got messages to save to the database!
WOOO HOOOO! so it's pretty ugly right now, i put the post message method
in the groupProject class, but i intend on cleaning it up a bit later.
the important part is that it works. I also added a getUserID method in
the UserAccount class. I actually didnt add anything to the message
class oddly enough, but we can fix that later.