You can browse messages now!
I'm still writing it all in grouproject.java, but I promise I'll chage
that after I get it all working. I dont believe I made any changes to
UserAccount.java or Message.java, but I uploaded them just to be safe.

ALSO: I'm having a hard time printing out the usernames for each message
so we still have to get all that sorted out (perhaps we could make that
separate-- like we give the user the option to select a message to view
more infor about it, like the username of the original poster, and maybe
an option to view their profile and add the message to like messages.)