Finished up
So users can do just about everything Dr. Anewalt asked for except for
viewing a user's profile description. Honestly, I don't know why this is
a problem, I'm gonna try to get that figured out before I submit it.
Again, the code is still very messy BUT IT WORKS, so I'm probably just
going to keep it the way it is until we present it on Tuesday.